Move Abstract Classes into Domain Folders
