+++ menrva/app/src/androidTest/cpp/jni/menrva_androidTest_EngineDebugger_Jni.cpp
@@ -23,7 +23,7 @@
 #include "../../../main/cpp/tools/ServiceLocator.h"
 #include "../../../main/cpp/ir/FirGenerator.h"
 #include "../../../main/cpp/effects/BassBoost.h"
-#include "../../../main/cpp/tools/WaveGenerator.h"
+#include "../../../main/cpp/ir/WaveGenerator.h"
 #include "../../../main/cpp/ModuleInterface.h"
 
 extern "C"
