+++ menrva/app/src/main/cpp/audio/AudioOutputBuffer.h
@@ -24,8 +24,8 @@
 #include "Sample.h"
 #include "AudioFormat.h"
 #include "../tools/ConversionBuffer.h"
-#include "../abstracts/LoggingBase.h"
-#include "../abstracts/AudioIOBufferBase.h"
+#include "../tools/LoggingBase.h"
+#include "AudioIOBufferBase.h"
 
 union audio_output_buffer_u {
     Buffer<int16_t>* PCM_16;
