+++ menrva/app/src/main/cpp/command_handlers/CommandHandlerBase.cpp
