+++ menrva/app/src/main/cpp/convolver/ConvolutionOperationsBase.h
