+++ menrva/app/src/main/cpp/effects/EffectBase.cpp
