+++ menrva/app/src/main/cpp/effects/MultiChannelEffectBase.cpp
