+++ menrva/app/src/main/cpp/effects/MultiChannelEffectBase.h
