+++ menrva/app/src/main/cpp/effects/SingleChannelEffectBase.cpp
