+++ menrva/app/src/main/cpp/engine/CommandHandlerMap.h
@@ -21,7 +21,7 @@
 
 #include <map>
 #include "../ModuleInterface.h"
-#include "../abstracts/CommandHandlerBase.h"
+#include "../command_handlers/CommandHandlerBase.h"
 
 typedef std::map<uint32_t, CommandHandlerBase*> handler_map;
