+++ menrva/app/src/main/cpp/fft/KissFftInterface.h
@@ -20,8 +20,8 @@
 #define MENRVA_KISSFFT_INTERFACE_H
 
 #include <map>
-#include "../abstracts/FftInterfaceBase.h"
-#include "../abstracts/LoggingBase.h"
+#include "FftInterfaceBase.h"
+#include "../tools/LoggingBase.h"
 #include "KissFftFunctions.h"
 
 typedef std::map<std::string, kissfft_plan_pair> KissFftPlanCache;
