+++ menrva/app/src/main/cpp/tools/LoggerBase.h
