+++ menrva/app/src/main/cpp/tools/LoggingBase.cpp
