+++ menrva/app/src/main/cpp/tools/LoggingBase.h
